Split TestFramework functionality in subclass for electionperf.py
