+++ scripts/TestFramework.py
@@ -248,92 +248,6 @@ class TestFramework(object):
                 server_id_ip = tolaunch.pop(0)[1]
                 self._start_server(server_command, server_id_ip)
 
-    def _same(self, lst):
-        """
-        Check if all elements in a list are the same.
-        """
-
-        return len(set(lst)) == 1
-    
-    def _await_stable_leader(self, after_term=0):
-        while True:
-            server_beliefs = {}
-
-            # Only the running servers are considered, whereas the self.server_ids_ips contains all
-            # servers in the cluster.
-            for server_id_ip, server_process in self.server_processes.items():
-                server_beliefs[server_id_ip] = {'leader_id_ip': None,
-                                                'term': None,
-                                                'wake': None}
-                b = server_beliefs[server_id_ip]
-
-                for line in open('debug/server_%d' % server_id_ip[0]):
-
-                    m = re.search('All hail leader (\d+) for term (\d+)', line)
-                    if m is not None:
-                        b['leader_id_ip'] = filter(
-                                lambda server_id_ip: server_id_ip[0] == int(m.group(1)), 
-                                self.server_ids_ips)[0]
-                        b['term'] = int(m.group(2))
-                        continue
-
-                    m = re.search('Now leader for term (\d+)', line)
-                    if m is not None:
-                        b['leader_id_ip'] = server_id_ip
-                        b['term'] = int(m.group(1))
-                        continue
-
-                    m = re.search('Running for election in term (\d+)', line)
-                    if m is not None:
-                        b['wake'] = int(m.group(1))
-
-            terms = [b['term'] for b in server_beliefs.values()]
-            leaders_ids_ips = [b['leader_id_ip'] for b in server_beliefs.values()]
-
-            if self._same(terms) and terms[0] > after_term:
-                assert self._same(leaders_ids_ips), server_beliefs
-
-                return {'leader_id_ip': leaders_ids_ips[0],
-                        'term': terms[0],
-                        'num_woken': sum([1 for b in server_beliefs.values() if b['wake'] > after_term])}
-            else:
-                time.sleep(.25)
-                self.sandbox.checkFailures()
-    
-    def election_performance(self, repeat=100):
-        num_terms = []
-        num_woken = []
-
-        for i in range(repeat):
-            old = self._await_stable_leader()
-            print('Server %d is the leader in term %d' % (old['leader_id_ip'][0], old['term']))
-
-            self._kill_server(old['leader_id_ip'])
-
-            new = self._await_stable_leader(after_term=old['term'])
-            print('Server %d is the leader in term %d' % (new['leader_id_ip'][0], new['term']))
-
-            self.sandbox.checkFailures()
-
-            num_terms.append(new['term'] - old['term'])
-            print('Took %d terms to elect a new leader' % (new['term'] - old['term']))
-            num_woken.append(new['num_woken'])
-            print('%d servers woke up' % (new['num_woken']))
-
-            self._start_server('build/LogCabin', old['leader_id_ip'])
-
-        num_terms.sort()
-        print('Num terms:', 
-            file=sys.stderr)
-        print('\n'.join(['%d: %d' % (i + 1, term) for (i, term) in enumerate(num_terms)]),
-            file=sys.stderr)
-
-        num_woken.sort()
-        print('Num woken:',
-            file=sys.stderr)
-        print('\n'.join(['%d: %d' % (i + 1, n) for (i, n) in enumerate(num_woken)]),
-            file=sys.stderr)
-
     def cleanup(self, debug=False):
         """
         Clean up the environment: configuration files, debug files and storage folders. Also,
